+++ packages/live-share/src/index.ts
@@ -3,8 +3,6 @@
  * Licensed under the Microsoft Live Share SDK License.
  */
 
-export * from "./DynamicObjectRegistry";
-export * from "./DynamicObjectManager";
 export * from "./HostTimestampProvider";
 export * from "./interfaces";
 export * from "./LiveEvent";
@@ -28,12 +26,12 @@ export * from "./LiveDataObject";
 export * from "./LiveShareRuntime";
 export * from "./AzureLiveShareClient";
 export * from "./AzureLiveShareHost";
-export * from "./BaseLiveShareClient";
 export * from "./LiveObjectSynchronizer";
 export * from "./errors";
-export * from "./smuggle";
-export {
-    TurboDirectory,
-    TurboDynamicObjects,
-    TurboStateMap,
-} from "./internals/consts";
+
+// TODO: our package structure is such that things that need to be exposed at all for other live share libs
+// are in the root of the directory even if they are "internal". move some to internal again?
+export * from "./internals/DynamicObjectRegistry";
+export * from "./internals/DynamicObjectManager";
+export * from "./internals/BaseLiveShareClient";
+export * from "./internals/smuggle";
