+++ packages/live-share/src/internals/DynamicObjectManager.ts
@@ -13,7 +13,7 @@ import {
 import { assert } from "@fluidframework/core-utils/internal";
 import { ConsensusRegisterCollection } from "@fluidframework/register-collection/internal";
 import { DynamicObjectRegistry } from "./DynamicObjectRegistry";
-import { LiveDataObject } from "./LiveDataObject";
+import { LiveDataObject } from "../LiveDataObject";
 
 // Register ConsensusRegisterCollection
 DynamicObjectRegistry.registerObjectClass(
