+++ packages/live-share/src/internals/DynamicObjectRegistry.ts
