+++ packages/live-share/src/internals/consts.ts
@@ -41,18 +41,3 @@ export const ObjectSynchronizerEvents = {
  * @hidden
  */
 export const LiveShareReportIssueLink = "https://aka.ms/teamsliveshare/issue";
-
-/**
- * @hidden
- */
-export const TurboDirectory = "TURBO_DIRECTORY";
-
-/**
- * @hidden
- */
-export const TurboStateMap = "TURBO_STATE_MAP";
-
-/**
- * @hidden
- */
-export const TurboDynamicObjects = "TURBO_DYNAMIC_OBJECTS";
